Fix old tests to work with new stuff
- New fast_overlap needs to convert labels to np.int32
- Test_tracking data was generated with the old default value for min_distance
  now min_distance is specified to be the old value.

+++ microutil/tests/test_tracking.py
@@ -9,7 +9,7 @@ from .utils import load_zarr, open_zarr
 
 def test_reindex_area():
     input = load_zarr('test-data/tracking/input').astype(int)[None, ...]
-    ds = xr.Dataset({'labels': xr.DataArray(input, dims=("S", "T", "Y", "X"))})
+    ds = xr.Dataset({'labels': xr.DataArray(input, dims=("S", "T", "Y", "X")).astype(np.int32)})
     expected = load_zarr('test-data/tracking/expected').astype(int)[None, ...]
     track(ds)
     assert np.allclose(expected, ds['labels'].values)
@@ -18,7 +18,7 @@ def test_reindex_area():
 # don't want to mess with test data so second test
 # so we can load in the different way
 def test_tracking2():
-    input = open_zarr('test-data/tracking2/input').load()
+    input = open_zarr('test-data/tracking2/input').load().astype(np.int32)
     expected = open_zarr('test-data/tracking2/expected').load()
     mu.track(input)
     assert np.allclose(input['labels'].values, expected['labels'].values)
